+++ public/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 - Checked metatag for filtering titles that have been checked.
   - Example: `checked:yes` would find everything that has been checked.
   - Example: `checked:no` would find everything that isn't checked.
+- Confirm when moving to a new page while having checked rows.
 
 ### Changed
 - `[Backend]` Updated Ion Auth.